% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_saf.R
\name{theme_saf}
\alias{theme_saf}
\title{theme_saf}
\usage{
theme_saf(base_size = 12, base_family = "Futura")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Font family.}
}
\value{
My ggplot2 theme (similar to theme_minimal/theme_bw)
}
\description{
My ggplot2 theme
}
